Fixes #203. Auto Refresh Not Working Properly
- added the possibility to turn on/off websockets in the player settings.
- TODO: fix it so turning off websockets, results in getting the messages only every time you hit a command. -> the old way.

+++ karchangame/src/main/java/mmud/rest/webentities/PrivateLog.java
@@ -19,13 +19,12 @@ package mmud.rest.webentities;
 import jakarta.xml.bind.annotation.XmlRootElement;
 
 /**
- *
  * @author maartenl
  */
 @XmlRootElement
 public class PrivateLog
 {
-    public int offset;
-    public int size;
-    public String log;
+  public long offset;
+  public int size;
+  public String log;
 }
